Remove redundant HTTPClientSpy as we already have one
