+++ EssentialDevelopper/EssentialDeveloperFrameworkTests/Feed API/LoadFeedFromRemoteUseCaseTests.swift
@@ -167,31 +167,4 @@ class LoadFeedFromRemoteUseCaseTests: XCTestCase {
         }
     }
     
-    class HTTPClientSpy: HTTPClient {
-        var messages = [(url: URL, completion:(HTTPClient.Result) -> ())]()
-        var cancelledURLs: [URL] = []
-        var requestedURLs: [URL] {
-            return messages.map { $0.url }
-        }
-        
-        func get(from url: URL, completion: @escaping (HTTPClient.Result) -> ()) -> HTTPClientTask {
-            messages.append((url, completion))
-            
-            return HTTPTask(cancelCompletion: { self.cancelledURLs.append(url) })
-        }
-        
-        func complete(with error: Error, at index: Int = 0) {
-            self.messages[index].completion(.failure(error))
-        }
-        
-        func completeWith(statusCode: Int, data: Data = Data(), at index: Int = 0) {
-            let httpResponse = HTTPURLResponse(
-                url: requestedURLs[index],
-                statusCode: statusCode,
-                httpVersion: nil,
-                headerFields: nil)!
-            
-            self.messages[index].completion(.success((httpResponse, data)))
-        }
-    }
 }
